UL-10 Updated account management views package
Moved ninja api instance into individual views module
due to requirements of ninjaAPI. This is to avoid
clashing API namespaces.
